No longer store workspace list with subdata.
1) Each workspace costs 64 bytes. Hence if we reserve 1Mb of the 16Mb
doc size that gives us 20k workspaces, which may not be enough for
popular objects, assuming KBase is successful. Reserving 5Mb = 100k
objects, which still might not be enough, and reduces subdata size by
33%.

2) It's become apparent that having 2 sources of truth in mongod is
dangerous unless they're both immutable. In this case, if a write or
delete is interrupted by mongo or the server going down, the workspaces
list could be inaccurate. Since mongo has no transactions there's not
really a reasonable way to deal with this other than forcing a single
source of truth.