Since it seems that the bizarro memory loss is nothing to do with my code...
...committing the complete big data test. Requires 3G to run, which
seems
reasonable since it generates ~1G characters, and each char is 2B in
java. With pointers, overhead and all the other junk in memory another
1G seems reasonable.

Bizarro memory loss = the JVM heap drops from 3G to ~2.8G after
retrieving the data from GridFS. However, this seems to be unrelated to
the code here since it can be replicated simply by storing 1M 1K strings
in a list.