Compare the writeTokens() method in JTS with Object.write(JG).
Writers passed to JsonTokenStream are wrapped with another writer that
deals with large strings. Passing in a JsonGenerator that wraps a writer
is the equivalent of UObject.write(OutputStream) except the second
wrapping doesn't occur. This second wrapping accounts for the 33%
performance difference between Object.write(OutputStream) and
Object.write(JsonGenerator).