Use configured user to validate user names via globus
The configured user should be a kbase_users admin to have visibility of
all users

Note that the static AuthService client is used in many locations
throughout the code; the configurable version is only used where
necessary. In order to mock an auth server in the future the static uses
need to be swapped for the configurable version (or the static class
needs to be configurable, which is possible).