Start removing subset calculate and storing code
Remove subset specific fields from SubsetAndMetadataNode and make
remaining tests pass

Note there was a weird interaction between the subsetting and metadata
extraction that caused the KB.MetadataT8.instance.1 test to return the
wrong answer and it was coded into the tests. Corrected the test to give
the right answer.