Switch to dependency injection for MongoWorkspaceDB init
Provide dependencies to MongoWorkspaceDB class in the constructor. 3
constructors with up to 8 arguments -> 1 constructor with 4 arguments. 

Makes it possible to change the configuration of the DB without changing
the MWDB class code - e.g. change the file backend used simply by
passing in a new BlobStore subclass. Same with the type validator (and
subsequently the type database, since validator construction is now
handled outside the MWDB class).