Encapsulate metadata & check key / value lengths.
I knew I should've encapsulated the meta maps when I coded it the first
time... this is what you get for being lazy.

All incoming metadata is encapsulated in a class that checks that all
keys and values are < 1000B (to prevent indexing failures in MongoDB).

The max metadata size check is now localized to the same class.

Outgoing metadata is encapsulated in a class that does no checking since
it's assumed if the metadata made it to the DB it's ok.

TODO: more tests & docs (see TODOs)