Remove skip, deprecate list_referencing_object_counts
Skip has all sorts of bad behavior associated with it which has been
discussed ad nauseum

lroc is not really useful generally - this type of thing should be done
in the nightly metrics, if at all. Also, the current plan for GC (if we
ever get around to it) would make the current implementation break, and
alternate implementations would be much slower.