Add a TypeProvider layer between the TypeValidator and TypeDB
Allows for alternatives to a direct typeDB connection when using the
type validator - for instance, might want to use in other code that
doesn't have access to the mongo instance. This allows writing a
TypeProvider that could call the workspace service to get type
information. Could also add caching in the provider layer.

Also remove instance variable for the TempFilesManager from Workspace.
Not needed.

All tests pass.