WIP on merging getReferencedObjects
Under the hood get objects, subobjects, provenance, and referenced
objects all are merged into getObjects at the WorkspaceMongoDB level.

Moved most of the getReferencedObjects logic into the Workspace class.

Stopped reporting when an object in a chain doesn't exist. That should
only happen when a user specifies an invalid chain. Just report that
there's no reference from the upstream object to the downstream object
in all cases.

TODO:
  add tests for reference chains where the references are not absolute
with happy and sad cases.
  Add refchain # and position for no such ref errors.
  see TODO REFS for future work after this merge is done.