Fix bug with register_typespec_copy & updated specs
Since allTypes was a list rather than a set, it wound up with two copies
of each type when the types in the new specfile were the same as the old
specfile. In the first time through the loop, the type was removed from
typesToSave, and in the second time added to typesToUnregister. Thus any
types that are in the new spec and the old spec will be unregistered.

Changing the list to a set fixed the bug - the tests now fail on a list,
but pass with a set.