Remove jongo overlay from TDB addRefs method
Adds an object mapper to replace the mapper in Jongo. Unlike the WS DB
provenance object, the objects in the TDB look simple enough that this
approach should work in a lot, if not all, cases.

Since the load on the TBD is very low and the TD objects are small, not
worried about any overhead from remapping the objects.