Add cache for admin role to workspace administration method
Since the workpace administration call is used by search and other
services to retrieve data, we cache the administration role of a user
for 5 minutes to reduce load on the admin role source (basically the
auth server) and speed up the calls. No point in fetching the user data
over and over every millisecond.