Remove SortCheckingTokenSequenceProvider
Functionality is a strict subset of IdRefTokenSequenceProvider. The
latter keeps track of the json location and jsonschema location, which
the former does not. Keeping track of the json location is required to
fix the sort checking bug.

There should be negligible performance hit for always using the
additional functionality since the schema and location data are
relatively small and in memory.

If SCTSP is kept, we'll have to add the location tracking functionality
and fix the bug in both places, so better to simplify and remove.