Fix sort checking bug on embedded maps
When a map was embedded within another map (even in an array) the last
key of the embedded map would be compared with the next key of the outer
map, which could cause the data to be erroneously marked as unsorted.
When exiting a map, set the previous map key correctly based on the
known location within the JSON structure.

The IdRefTokenSequenceProviderTest tests sortedEmbeddedMap and
sortedEmbeddedMapInList failed prior to the fix. The other tests were
added after the fix was in place.