Presigned URL workaround working.
Switched to Apache http client while trying to figure out how to respond
to minio errors. Looks like small uploads can read the error but for
large uploads the connection is reset. Not sure how to deal with this
otherwise.

Going to go with the http client, much easier to work with.