Fix potential but unlikely failure mode
It's possible that a put object call to minio could result in a 3XX if
the reverse proxy config is broken or something along those lines. In
that case, no exception would be thrown but the object would not be
saved. Lock things down to be as conservative as possible when checking
the return code.