Remove unused cruft
For some reason getting the list of workspaces to which a user had
access allows specifying multiple users, which confused the hell out of
me. Turns out this feature and a couple of helper methods are unused, so
in the garbage they go.

Also some other minor cruft cleanup and force a cursor close.