Required workspaces in list_object, refactor parameters
This commit turned out to be really big. Sorry about that.

It does two major things, which wound up having a number of knock on
effects:

1. It refactors the list object parameters class to use a builder rather
than a zillion constructors.
  1. It also modernized the class a bit, using Optional and Instant
instead of nulls and Dates.
2. It requires between 1 and 10000 workspaces as inputs when listing
objects. Type only lists are dangerous because a large number of
workspaces (potentially >> 10000) may need to be examined as part of the
operation.

That has some knock on effects:
1. the minimum permission parameter for list objects is now useless,
since a workspace list is required, so the user should just filter
there.
2. likewise, the exclude global workspaces command is also useless.