Switch from the Mongo 2.X api to the 3.0 API.
Not really any way to do the switch piecemeal - once you switch from
getDB() to getDatabase() you're on the new API and have to make changes
everywhere.

Lots and lots of potential refactoring was unearthed while making these
changes, but was only done where not doing it would have been more work
in order to minimize the changed line count, which is already large.