Add method to add data to WorkspaceObjectData builder
Next step is to return no data from the MWDB.getObjects method and
use this method to populate the builder with data. Then the data size
calculation can be done in the Workspace class and the
ByteArrayFileCacheManager can be initialized up front to save to memory
or disk, getting rid of the thread unsafe dynamic reallocation code.