Rewrite ByteArrayFileCacheManager for thread safety
In the current implementation, BAFCM keeps a running tally of the size
of data in memory and when that size is exceeded it dumps the current
file to disk. It updates the tally after each file is completed, which
assumes only one file is being processed at a time.

This is not safe for the case where data creation is parallelized and
multiple data objects are streaming into memory simultaneously. This
commit rewrites BAFCM to simply put all data into memory or disk,
depending on the wishes of the creator of the BAFCM. The Workspace class
is the creator, which chooses to use disk storage if the size of the
incoming data + potential subsets is greater than the configured maximum
allowed.

This does have a drawback that if you have large data but small subsets,
the data may be stored on disk when it might have fit in memory in the
prior implementation. However, this is a very specific case for a very
specific amount of data that is close to the in memory limit. If there
are no subsets, the subsets are close to the size of the objects, the
data exceeds the in memory limit, or the data is not close to the in
memory limit there's no effective difference.

An alternative implementation could keep track of all simultaneous
incoming data byte by byte, but that seems like a great deal of
complication and source of parallelization bugs to fix a relatively
small proportion of cases.

TODO: Javadoc for BAFCM